svg/geometry/parsing/cx-computed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="8000"><title>SVG Geometry Properties: getComputedStyle().cx</title><metadata><h:link href="https://svgwg.org/svg2-draft/geometry.html#CX" rel="help"/></metadata><style>#target{font-size:40px}</style><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/computed-testcommon.js"/><script><![CDATA[

test_computed_value("cx", "-10px");
test_computed_value("cx", "0.5em", "20px");
test_computed_value("cx", "calc(10px + 0.5em)", "30px");
test_computed_value("cx", "calc(10px - 0.5em)", "-10px");
test_computed_value("cx", "40%");
test_computed_value("cx", "calc(50% + 60px)");

  ]]></script></svg>
